share jvm/native stdin/out/err + resource code
