+++ io/native/src/main/scala/fs2/io/ioplatform.scala
@@ -33,60 +33,4 @@ import java.nio.charset.Charset
 import java.nio.charset.StandardCharsets
 import scala.reflect.ClassTag
 
-private[fs2] trait ioplatform {
-  type InterruptedIOException = java.io.InterruptedIOException
-  type ClosedChannelException = java.nio.channels.ClosedChannelException
-
-  //
-  // STDIN/STDOUT Helpers
-
-  /** Stream of bytes read asynchronously from standard input. */
-  def stdin[F[_]: Sync](bufSize: Int): Stream[F, Byte] =
-    readInputStream(Sync[F].blocking(System.in), bufSize, false)
-
-  /** Pipe of bytes that writes emitted values to standard output asynchronously. */
-  def stdout[F[_]: Sync]: Pipe[F, Byte, Nothing] =
-    writeOutputStream(Sync[F].blocking(System.out), false)
-
-  /** Pipe of bytes that writes emitted values to standard error asynchronously. */
-  def stderr[F[_]: Sync]: Pipe[F, Byte, Nothing] =
-    writeOutputStream(Sync[F].blocking(System.err), false)
-
-  /** Writes this stream to standard output asynchronously, converting each element to
-    * a sequence of bytes via `Show` and the given `Charset`.
-    */
-  def stdoutLines[F[_]: Sync, O: Show](
-      charset: Charset = StandardCharsets.UTF_8
-  ): Pipe[F, O, Nothing] =
-    _.map(_.show).through(text.encode(charset)).through(stdout)
-
-  /** Stream of `String` read asynchronously from standard input decoded in UTF-8. */
-  def stdinUtf8[F[_]: Sync](bufSize: Int): Stream[F, String] =
-    stdin(bufSize).through(text.utf8.decode)
-
-  /** Stream of bytes read asynchronously from the specified resource relative to the class `C`.
-    * @see [[readClassLoaderResource]] for a resource relative to a classloader.
-    */
-  def readClassResource[F[_], C](
-      name: String,
-      chunkSize: Int = 64 * 1024
-  )(implicit F: Sync[F], ct: ClassTag[C]): Stream[F, Byte] =
-    Stream.eval(F.blocking(Option(ct.runtimeClass.getResourceAsStream(name)))).flatMap {
-      case Some(resource) => io.readInputStream(resource.pure, chunkSize)
-      case None           => Stream.raiseError(new IOException(s"Resource $name not found"))
-    }
-
-  /** Stream of bytes read asynchronously from the specified classloader resource.
-    * @see [[readClassResource]] for a resource relative to a class.
-    */
-  def readClassLoaderResource[F[_]](
-      name: String,
-      chunkSize: Int = 64 * 1024,
-      classLoader: ClassLoader = getClass().getClassLoader()
-  )(implicit F: Sync[F]): Stream[F, Byte] =
-    Stream.eval(F.blocking(Option(classLoader.getResourceAsStream(name)))).flatMap {
-      case Some(resource) => io.readInputStream(resource.pure, chunkSize)
-      case None           => Stream.raiseError(new IOException(s"Resource $name not found"))
-    }
-
-}
+private[fs2] trait ioplatform extends iojvmnative
